+++ package.json
@@ -55,6 +55,7 @@
         "@jupyter-widgets/base": "^1.1.10 || ^2.0.0 || ^3.0.0 || ^4.0.0",
         "@mantine/core": "^5.10.0",
         "@mantine/hooks": "^5.10.0",
+        "@tabler/icons-react": "^2.14.0",
         "react": "^18.2.0",
         "react-dom": "^17.0.2",
         "react-icons": "^4.7.1",
